+++ tests/test_calibration.py
@@ -6,31 +6,7 @@ import os
 
 import pytest
 
-from openfisca_core.taxbenefitsystems import TaxBenefitSystem  # type: ignore
-from openfisca_france_data import france_data_tax_benefit_system  # type: ignore
-from openfisca_france_data.erfs.scenario import ErfsSurveyScenario  # type: ignore
 from openfisca_survey_manager.calibration import Calibration  # type: ignore
-
-
-@pytest.fixture
-def tax_benefit_system() -> TaxBenefitSystem:
-    return france_data_tax_benefit_system
-
-
-@pytest.fixture
-def year() -> int:
-    return 2009
-
-
-@pytest.fixture
-def survey_scenario(
-        tax_benefit_system: TaxBenefitSystem,
-        year: int,
-        ) -> ErfsSurveyScenario:
-    return ErfsSurveyScenario.create(
-        tax_benefit_system = tax_benefit_system,
-        year = year,
-        )
 
 
 @pytest.fixture
@@ -38,8 +14,20 @@ def location() -> str:
     return pkg_resources.get_distribution("openfisca-france-data").location
 
 
-@pytest.mark.skip(reason = "assert data is not None")
-def test_calibration(survey_scenario, year, location):
+@pytest.mark.skip(
+    reason = "'Calibration' object has no attribute 'set_inputs_margins_from_file'",
+    )
+def test_calibration(survey_scenario, fake_input_data, location, year: int = 2009):
+    # On ititialise le survey scenario
+    survey_scenario = survey_scenario(year)
+
+    # On charge les données
+    input_data = fake_input_data(year)
+
+    # On initialise le survey scenario
+    survey_scenario.init_from_data(data = dict(input_data_frame = input_data))
+
+    # On fait la calibration
     calibration = Calibration(survey_scenario, period = year)
     calibration.parameters["method"] = "linear"
     calibration.total_population = calibration.initial_total_population * 1.123
